Rewrite EstimationTask as ConcurrentTaskG<EstimationResult> Put oneaudit in separate package.
