+++ core/src/main/kotlin/org/cryptobiotic/rlauxe/sampling/EstimationResult.kt
@@ -1,0 +1,7 @@
+package org.cryptobiotic.rlauxe.sampling
+
+data class EstimationResult(
+    val task: SimulateSampleSizeTask,
+    val repeatedResult: RunTestRepeatedResult,
+    val failed: Boolean
+)
