+++ core/src/main/kotlin/org/cryptobiotic/rlauxe/sampling/RunRepeated.kt
@@ -5,38 +5,6 @@ import org.cryptobiotic.rlauxe.core.TestH0Status
 import org.cryptobiotic.rlauxe.util.*
 import kotlin.math.sqrt
 
-data class RunTestRepeatedResult(
-    val testParameters: Map<String, Double>, // various parameters, depends on the test
-    val Nc: Int,                  // population size (eg number of ballots)
-    val totalSamplesNeeded: Int, // total number of samples needed in nsuccess trials
-    val nsuccess: Int,           // number of successful trials
-    val ntrials: Int,            // total number of trials
-    val variance: Double,        // variance over ntrials of samples needed
-    val percentHist: Deciles? = null, // histogram of successful sample size as percentage of N, count trials in 10% bins
-    val status: Map<TestH0Status, Int>? = null, // count of the trial status
-    val sampleCount: List<Int> = emptyList(),
-    val errorRate: List<Double> = emptyList(), // error rates percent
-    val margin: Double?,
-) {
-
-    fun successPct(): Double = 100.0 * nsuccess / (if (ntrials == 0) 1 else ntrials)
-    fun failPct(): Double  = 100.0 * (ntrials - nsuccess) / (if (ntrials == 0) 1 else ntrials)
-    fun avgSamplesNeeded(): Int  = totalSamplesNeeded / (if (nsuccess == 0) 1 else nsuccess)
-    fun pctSamplesNeeded(): Double  = 100.0 * avgSamplesNeeded().toDouble() / (if (Nc == 0) 1 else Nc)
-    fun errorRates() = buildString { errorRate.forEach{ append("${df(it)},") } }
-
-    override fun toString() = buildString {
-        appendLine("RunTestRepeatedResult: testParameters=$testParameters Nc=$Nc successPct=${successPct()} in ntrials=$ntrials")
-        append("  $nsuccess successful trials: avgSamplesNeeded=${avgSamplesNeeded()} stddev=${sqrt(variance)}")
-        if (percentHist != null) appendLine("  cumulPct:${percentHist.cumulPct()}") else appendLine()
-        if (status != null) appendLine("  status:${status}")
-    }
-
-    fun findQuantile(quantile: Double): Int {
-        return quantile(sampleCount, quantile)
-    }
-}
-
 // single threaded
 // runs RiskTestingFn repeatedly, drawSample.reset() called for each trial.
 fun runTestRepeated(
@@ -97,9 +65,34 @@ fun runTestRepeated(
         ntrials=ntrials, variance, percentHist, statusMap, sampleCounts, errorCounts.map { 100.0 * it / ntrials}, margin = margin)
 }
 
-fun ceilDiv(numerator: Int, denominator: Int): Int {
-    val frac = numerator.toDouble() / denominator
-    val fracFloor = frac.toInt()
-    val fracCeil = if (frac == fracFloor.toDouble()) fracFloor else fracFloor + 1
-    return fracCeil
+data class RunTestRepeatedResult(
+    val testParameters: Map<String, Double>, // various parameters, depends on the test
+    val Nc: Int,                  // population size (eg number of ballots)
+    val totalSamplesNeeded: Int, // total number of samples needed in nsuccess trials
+    val nsuccess: Int,           // number of successful trials
+    val ntrials: Int,            // total number of trials
+    val variance: Double,        // variance over ntrials of samples needed
+    val percentHist: Deciles? = null, // histogram of successful sample size as percentage of N, count trials in 10% bins
+    val status: Map<TestH0Status, Int>? = null, // count of the trial status
+    val sampleCount: List<Int> = emptyList(),
+    val errorRate: List<Double> = emptyList(), // error rates percent
+    val margin: Double?,
+) {
+
+    fun successPct(): Double = 100.0 * nsuccess / (if (ntrials == 0) 1 else ntrials)
+    fun failPct(): Double  = 100.0 * (ntrials - nsuccess) / (if (ntrials == 0) 1 else ntrials)
+    fun avgSamplesNeeded(): Int  = totalSamplesNeeded / (if (nsuccess == 0) 1 else nsuccess)
+    fun pctSamplesNeeded(): Double  = 100.0 * avgSamplesNeeded().toDouble() / (if (Nc == 0) 1 else Nc)
+    fun errorRates() = buildString { errorRate.forEach{ append("${df(it)},") } }
+
+    override fun toString() = buildString {
+        appendLine("RunTestRepeatedResult: testParameters=$testParameters Nc=$Nc successPct=${successPct()} in ntrials=$ntrials")
+        append("  $nsuccess successful trials: avgSamplesNeeded=${avgSamplesNeeded()} stddev=${sqrt(variance)}")
+        if (percentHist != null) appendLine("  cumulPct:${percentHist.cumulPct()}") else appendLine()
+        if (status != null) appendLine("  status:${status}")
+    }
+
+    fun findQuantile(quantile: Double): Int {
+        return quantile(sampleCount, quantile)
+    }
 }
