+++ core/src/main/kotlin/org/cryptobiotic/rlauxe/util/Utils.kt
@@ -18,6 +18,13 @@ fun doubleIsClose(a: Double, b: Double, rtol: Double=1.0e-5, atol:Double=1.0e-8)
 
 fun roundToInt(x: Double) = round(x).toInt()
 
+fun ceilDiv(numerator: Int, denominator: Int): Int {
+    val frac = numerator.toDouble() / denominator
+    val fracFloor = frac.toInt()
+    val fracCeil = if (frac == fracFloor.toDouble()) fracFloor else fracFloor + 1
+    return fracCeil
+}
+
 fun listToMap(vararg names: String): Map<String, Int> {
     return names.mapIndexed { idx, value -> value to idx }.toMap()
 }
