+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/concur/RepeatedTaskRunner.kt
@@ -1,11 +1,11 @@
 
 package org.cryptobiotic.rlauxe.concur
 
-import org.cryptobiotic.rlauxe.core.TestH0Status
 import org.cryptobiotic.rlauxe.workflow.WorkflowResult
 import org.cryptobiotic.rlauxe.workflow.WorkflowTaskGenerator
 
-class RepeatedTaskRunner (val nruns: Int, val taskGenerator: WorkflowTaskGenerator): ConcurrentTaskG<List<WorkflowResult>> {
+class RepeatedTaskRunner (val nruns: Int, val taskGenerator: WorkflowTaskGenerator):
+    ConcurrentTaskG<List<WorkflowResult>> {
 
     override fun name(): String = "Repeated-${taskGenerator.name()}"
