+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/workflow/GenAuditsWithPhantomsPlots.kt
@@ -38,7 +38,7 @@ class GenAuditsWithPhantomsPlots {
         }
 
         // run tasks concurrently and average the results
-        val results: List<WorkflowResult> = runRepeatedTaskAndAverage(tasks)
+        val results: List<WorkflowResult> = runRepeatedWorkflowsAndAverage(tasks)
         println(stopwatch.took())
 
         val writer = WorkflowResultsIO("$dirName/${name}.cvs")
