+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/workflow/GenAuditsWithUndervotesPlots.kt
@@ -38,7 +38,7 @@ class GenAuditsWithUndervotesPlots {
         }
 
         // run tasks concurrently and average the results
-        val results: List<WorkflowResult> = runRepeatedTaskAndAverage(tasks)
+        val results: List<WorkflowResult> = runRepeatedWorkflowsAndAverage(tasks)
         println(stopwatch.took())
