+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/workflow/WorkflowTasks.kt
@@ -2,17 +2,12 @@ package org.cryptobiotic.rlauxe.workflow
 
 import org.cryptobiotic.rlauxe.concur.ConcurrentTaskG
 import org.cryptobiotic.rlauxe.concur.ConcurrentTaskRunnerG
-import org.cryptobiotic.rlauxe.concur.RepeatedTaskRunner
 import org.cryptobiotic.rlauxe.core.Cvr
 import org.cryptobiotic.rlauxe.core.TestH0Status
 import org.cryptobiotic.rlauxe.oneaudit.makeContestOA
-import org.cryptobiotic.rlauxe.rlaplots.WorkflowResultsIO
-import org.cryptobiotic.rlauxe.rlaplots.WorkflowResultsPlotter
 import org.cryptobiotic.rlauxe.sampling.ContestSimulation
 import org.cryptobiotic.rlauxe.sampling.makeFuzzedCvrsFrom
-import org.cryptobiotic.rlauxe.util.df
 import kotlin.random.Random
-import kotlin.test.Test
 
 private val quiet = true
 
@@ -145,7 +140,7 @@ class WorkflowTask(
     }
 }
 
-fun runRepeatedTaskAndAverage(tasks: List<ConcurrentTaskG<List<WorkflowResult>>>): List<WorkflowResult> {
+fun runRepeatedWorkflowsAndAverage(tasks: List<ConcurrentTaskG<List<WorkflowResult>>>): List<WorkflowResult> {
     val rresults: List<List<WorkflowResult>> = ConcurrentTaskRunnerG<List<WorkflowResult>>().run(tasks)
     val results: List<WorkflowResult> = rresults.map { avgWorkflowResult(it) }
     return results
@@ -188,90 +183,3 @@ fun avgWorkflowResult(runs: List<WorkflowResult>): WorkflowResult {
     }
 }
 
-class TestGenWorkflowTasks {
-    val nruns = 10
-
-    @Test
-    fun genPollingWorkflowMarginPlots() {
-        val N = 50000
-        val margins = listOf(.02, .03, .04, .05, .06, .07, .08, .09, .10)
-
-        val tasks = mutableListOf<ConcurrentTaskG<List<WorkflowResult>>>()
-        margins.forEach { margin ->
-            val workflowGenerator = PollingWorkflowTaskGenerator(N, margin, 0.0, 0.0, 0.0,
-                mapOf("nruns" to nruns.toDouble()))
-            tasks.add(RepeatedTaskRunner(nruns, workflowGenerator))
-        }
-
-        // run tasks concurrently and average the results
-        val results: List<WorkflowResult> = runRepeatedTaskAndAverage(tasks)
-
-        val dirName = "/home/stormy/temp/testWorkflowTasks"
-        val filename = "PollingMarginRepeated"
-        val writer = WorkflowResultsIO("$dirName/${filename}.cvs")
-        writer.writeResults(results)
-
-        val plotter = WorkflowResultsPlotter(dirName, filename)
-        plotter.showSampleSizesVsMargin(results, "category") { "all" }
-    }
-
-    @Test
-    fun genClcaWorkflowMarginPlots() {
-        val N = 50000
-        val margins = listOf(.02, .03, .04, .05, .06, .07, .08, .09, .10)
-
-        val tasks = mutableListOf<ConcurrentTaskG<List<WorkflowResult>>>()
-        margins.forEach { margin ->
-            val workflowGenerator = ClcaWorkflowTaskGenerator(N, margin, 0.0, 0.0, 0.0,
-                mapOf("nruns" to nruns.toDouble()))
-            tasks.add(RepeatedTaskRunner(nruns, workflowGenerator))
-        }
-
-        // run tasks concurrently and average the results
-        val results: List<WorkflowResult> = runRepeatedTaskAndAverage(tasks)
-
-        val dirName = "/home/stormy/temp/testWorkflowTasks"
-        val filename = "ClcaMargin"
-        val writer = WorkflowResultsIO("$dirName/${filename}.cvs")
-        writer.writeResults(results)
-
-        val plotter = WorkflowResultsPlotter(dirName, filename)
-        plotter.showSampleSizesVsMargin(results, "category") { "all" }
-    }
-
-    @Test
-    fun genOneAuditWorkflowMarginPlots() {
-        val N = 50000
-        val cvrPercents = listOf(.2, .4, .6, .8, .9, .95, .99)
-        val margins = listOf(.02, .03, .04, .05, .06, .07, .08, .09, .10)
-
-        val auditConfig = AuditConfig(
-            AuditType.ONEAUDIT,
-            hasStyles = true,
-            seed = Random.nextLong(),
-            quantile = .80,
-            ntrials = 10
-        )
-        println("N=${N} ntrials=${auditConfig.ntrials} fuzzPct = ${auditConfig.fuzzPct}")
-
-        val tasks = mutableListOf<ConcurrentTaskG<List<WorkflowResult>>>()
-        cvrPercents.forEach { cvrPercent ->
-            margins.forEach { margin ->
-                val workflowGenerator = OneAuditWorkflowTaskGenerator(N, margin, 0.0, 0.0, cvrPercent, 0.0,
-                    mapOf("nruns" to nruns.toDouble()))
-                tasks.add(RepeatedTaskRunner(nruns, workflowGenerator))
-            }
-        }
-
-        // run tasks concurrently and average the results
-        val results: List<WorkflowResult> = runRepeatedTaskAndAverage(tasks)
-
-        val dirName = "/home/stormy/temp/testWorkflowTasks"
-        val filename = "OneAuditMarginRepeated"
-        val writer = WorkflowResultsIO("$dirName/${filename}.cvs")
-        writer.writeResults(results)
-
-        val plotter = WorkflowResultsPlotter(dirName, filename)
-        plotter.showSampleSizesVsMargin(results, "cvrPercent") { df(it.parameters["cvrPercent"]!!.toDouble()) }
-    }
-}
